requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
streamlit==1.45.0
sqlalchemy==2.0.40
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
faiss-cpu==1.11.0
openai==1.77.0
httpx==0.28.1
numpy==2.2.5
PyPDF2==3.0.1
PyMuPDF==1.26.0
Pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
pytest==8.3.5
